project/requirements.txt
flask
openai
pytesseract
pillow
numpy